+++ cpp/src/gandiva/precompiled/arithmetic_ops_test.cc
@@ -21,14 +21,10 @@
 namespace gandiva {
 
 TEST(TestArithmeticOps, TestIsDistinctFrom) {
-  EXPECT_EQ(is_distinct_from_timestamp_timestamp((int64)1000, true, (int64)1000, false),
-            true);
-  EXPECT_EQ(is_distinct_from_timestamp_timestamp((int64)1000, false, (int64)1000, true),
-            true);
-  EXPECT_EQ(is_distinct_from_timestamp_timestamp((int64)1000, false, (int64)1000, false),
-            false);
-  EXPECT_EQ(is_distinct_from_timestamp_timestamp((int64)1000, true, (int64)1000, true),
-            false);
+  EXPECT_EQ(is_distinct_from_timestamp_timestamp(1000, true, 1000, false), true);
+  EXPECT_EQ(is_distinct_from_timestamp_timestamp(1000, false, 1000, true), true);
+  EXPECT_EQ(is_distinct_from_timestamp_timestamp(1000, false, 1000, false), false);
+  EXPECT_EQ(is_distinct_from_timestamp_timestamp(1000, true, 1000, true), false);
 
   EXPECT_EQ(is_not_distinct_from_int32_int32(1000, true, 1000, false), false);
   EXPECT_EQ(is_not_distinct_from_int32_int32(1000, false, 1000, true), false);
